Gil - CR

+++ src/main/java/frc/constants/RobotHeadingEstimatorConstants.java
@@ -5,7 +5,7 @@ public class RobotHeadingEstimatorConstants {
 
 	public static final double POSE_BUFFER_SIZE_SECONDS = 2.0;
 
-	public static final double AMOUNT_OF_SOURCE_TYPES = 2.0;
+	public static final int AMOUNT_OF_SOURCE_TYPES = 2;
 
 	public static final double DEFAULT_GYRO_STANDARD_DEVIATION = 0.0001;
